% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflattice.R
\name{admm_configuration}
\alias{admm_configuration}
\title{Configuration parameters for \code{\link[=etfgrid]{etfgrid()}}}
\usage{
admm_configuration(
  lambda_min = -1,
  lambda_max = -1,
  mu = -1,
  mu_adjust = -1,
  rho = -1,
  rho_adjust = -1,
  maxiter = 100L,
  tolerance = 0.001,
  verbose = 0L
)
}
\arguments{
\item{lambda_min}{Smallest value for the lambda sequence. Negative values
(the default) result in automatic selection  (1e-4 * lambda_max)}

\item{lambda_max}{Largest value for the lambda sequence. Negative values
(the default) result in automatic selection (infinity norm of \code{Dy} divided
by \code{n})}

\item{mu}{ADMM multiplier on the proximity part of the ADMM penalty (proximity to
the previous iterate). Expert use only.}

\item{mu_adjust}{Adjustment multiplier for \code{mu} (enforces proximity between
the current and previous value). Expert use only.}

\item{rho}{ADMM multiplier on the primal/dual part of the ADMM penalty.}

\item{rho_adjust}{Adjustment multiplier for \code{rho}.}

\item{maxiter}{Maximum number of ADMM iterates.}

\item{tolerance}{Tolerance for stopping the iteration (before maxiter).}

\item{verbose}{Controls the printing of progress. Larger values result in
more printing.}
}
\value{
a list.
}
\description{
This function is not intended for direct usage. Passing these arguments
to \code{\link[=etfgrid]{etfgrid()}} controlls the algorithmic implementation.
}
