% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_map}
\alias{canada_map}
\title{Spatial polygon map of Canadian Provinces}
\format{
map of class SpatialPolygons
}
\usage{
canada_map
}
\description{
Canadian provincial borders are rather complex (many islands and inlets).
The official map from StatCan is around 20 Mb and challenging to use to make
quick and dirty maps. This dataset fills the gap. It's a lower resolution
version.
}
\examples{
library(ggplot2)
ggplot() + 
  geom_path(data=canada_map,aes(x=long,y=lat,group=group),
  color="darkgreen", size = .5) +
  coord_map(projection = 'lambert', parameters=c(49,77)) +
  theme_bw()
}
\keyword{datasets}
