% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflattice.R
\name{etfgrid}
\alias{etfgrid}
\title{Exponential family trend filtering over grids}
\usage{
etfgrid(
  y,
  korder = rep(0L, d),
  exp_fam = c("gauss_mean", "gauss_var", "poisson", "exponential", "binomial"),
  dims = dim(y),
  wrap = rep(FALSE, d),
  lambda = numeric(0),
  n_solutions = 50L,
  penalize_nullspace = FALSE,
  nullspace_mult = 0.1,
  return_mean_param = TRUE,
  estimate_df = FALSE,
  ...
)
}
\arguments{
\item{y}{vector or array of observations}

\item{korder}{vector of trend filter orders, non-negative integers}

\item{exp_fam}{the name of the distribution/parameter, one of}

\item{dims}{vector of dimensions for the grid, can be specified if y is not
an array but a vector of length (n1 x n2 x ... nd)}

\item{wrap}{logical vector indicating if any dimensions are wrapped (like
a cylinder, say)}

\item{lambda}{non-negative, decreasing sequence of penalty parameters, will
be determined automatically if omitted}

\item{n_solutions}{the number of solutions to use if lambda is omitted}

\item{penalize_nullspace}{should we include a penalty on the nullspace
of the TF operator}

\item{nullspace_mult}{the nullspace penalty is lambda scaled by this value}

\item{return_mean_param}{logical, is the returned estimate the natural
parameter (TRUE returns the mean parameter)}

\item{estimate_df}{logical, should we calculate the degrees of freedom.
This can potentially be slow. It is very
fast for \code{gaussian_mean} without a nullspace penalty and slower for other
options. Larger grids increase
the computing time very quickly.}

\item{...}{additional arguments to configure the ADMM algorithm.
See \code{\link[=admm_configuration]{admm_configuration()}}}
}
\value{
an object of class \link{etfgrid}
}
\description{
Exponential family trend filtering over grids
}
