% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-estimation.R
\name{kl_estimate}
\alias{kl_estimate}
\title{Unbiased estimation of the KL loss}
\usage{
kl_estimate(object)
}
\arguments{
\item{object}{an object of class \code{etfgrid} as returned by a call to \code{\link[=etfgrid]{etfgrid()}}.
The object must include an estimate of the degrees of freedom, except in
the case of Poisson. For all other cases one must use
\code{etfgrid(..., estimate_df = TRUE)}.}
}
\value{
A data frame with two or three columns. The first contains the values
of \code{lambda} used by \code{\link[=etfgrid]{etfgrid()}} the second contains unbiased estimates of
the expected Kullback-Leibler loss of the fit. Models with lower values
are preferred. When the family is other than Poisson, the estimated
degrees is the third column in the output. The returned object is an S3
object with plot and print methods
}
\description{
Unbiased estimation of the KL loss
}
\examples{
y <- rnorm(100, c(rep(0,30), rep(5,40), rep(0,30)))
out <- etfgrid(y, estimate_df = TRUE, lambda_max = 10)
risk <- kl_estimate(out)
}
