% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty_matrix}
\alias{penalty_matrix}
\title{Trend filter penalty matrix}
\usage{
penalty_matrix(
  side_lengths,
  orders = rep(0L, length(side_lengths)),
  wrap = rep(FALSE, length(side_lengths))
)
}
\arguments{
\item{side_lengths}{vector of side lengths for the grid}

\item{orders}{vector of orders of the trend filter penalty (denoted \code{k}).}

\item{wrap}{vector indicating any dimensions to "wrap". That is whether
the left side wraps around to tough the right side (like a cylinder).}
}
\value{
a sparse matrix. See \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}}.
}
\description{
Trend filter penalty matrix
}
\examples{
penalty_matrix(c(4,5), c(1,0))
}
