% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_temperatures}
\alias{world_temperatures}
\title{Temperatures for the northern hemisphere in 2010}
\format{
A 2880 row by 53 column tibble. Columns are \code{longitude}, \code{latitude}
(duplicated as necessary) and the 51 complete weeks in 2010. This represents
80 longitudes (0 - 360 by 4) and 36 latitudes (0 - 80 by 2).
}
\source{
This is a subset of subsampled ERA-20C data available at \url{https://www.ecmwf.int/en/forecasts/dataset/ecmwf-reanalysis-20th-century}
}
\usage{
world_temperatures
}
\description{
This data represents 4-degree x 2-degree gridded temperature for the
northern hemisphere in 2010. Each week averages the daily data from the
original source.
}
\keyword{datasets}
